separated tests into another directory
